requirements.txt
requests==2.32.5
python-dotenv==1.2.1
pandas==2.3.3
psycopg2-binary==2.9.11
hyperliquid-python-sdk==0.22.0
eth-account==0.13.7
yfinance==0.2.54
